requirements.txt
Django~=3.2.5
python-slugify==5.0.2
Pillow==8.3.1
django-braces==1.14.0
django-embed-video==1.4.0